Update team details page - edit title
This commit makes it possible for a user to edit the title of a team
from the team details page.

Adds a button (styled like link) that, when clicked, will reveal
a form in place of the page header. The user can enter changes to the
team's title and click save or cancel to hide the form.

Adds patchTeam method to teams controller.

Adds PATH /api/teams/:id route to handle request to update the
given values for a team.

Closes #42