Format urls in notes on team details view
Any urls found within the notes on the team-details view are replaced
with links (<a>) elements that navigate to the url by opening a new
tab.

This is implemented in team-details.js. The formatting happens once
the DOM content has loaded by finding elements with the "linkify" class
and replacing all urls with <a> tags using the url for the href and
text content of the <a>.

Closes #45