Preserve white space in notes on the details page
Updates common.css and team-details view in order to render the notes
with any line breaks for paragraphs, tabs, and other user entered
white space preserved for better readability in longer notes.

Adds a .pre-wrap class selector to common.css and assigned the
"pre-wrap" class to the paragraph elements containing on the note text
on the team-details page.

Closes #46