Update createNote mutation - require auth
Update createNote resolver to return {success: false} when the given
teamId arg does not match a team belonging to the authenticated user.
Gets the logged in user from context.

Update Teams datasource to add user to the criteria for findAndUpdate.
Returns null if no team found matching the userId and teamId params.

Add test for require auth when creating a note to api.test.