Update the updateTeam mutation - require auth
Updates the updateTeam mutation so that a user may only update a team
which belongs to the user.

Update the resolver to pass a userId parameter for the user id in the
auth token to the update method of the teams datasource.

Update the update method of the teams datasource to accept a userId
property in the first parameter and use that one performing a
findOneAndUpdate query on the Team model.